requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
